third_party/blink/web_tests/platform/generic/svg/dom/operatorAttribute-expected.txt

This test checks that SVG operator attribute exists while _operator doesn't

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS feMorphologyElement.operator.baseVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE
PASS feMorphologyElement.operator.baseVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_ERODE
PASS feMorphologyElement._operator is undefined.
PASS successfullyParsed is true

TEST COMPLETE
